requirements.txt
chromadb
python-dotenv
fitz